Fix tab movement and duplication
This commit fixes broken local buffer lists for moved and duplicated
frames.  It removes the `bufferlo-desktop-support' custom variable.
Desktop support for local frames is now always enabled when
`bufferlo-mode' is active (was the default before), as deactivation
conflicts with tab movement, and it makes no sense to artificially
disable it.