Rename bookmark "session" to "set"
Once in a project's lifetime ability to rename a high-level feature so
here we go.

This also renames the bookmark handler from
bufferlo--bookmark-session-handler to bufferlo--bookmark-set-handler
so any existing bookmark files need to be hand edited to reflect the
change in handler name.

References to session functions associated with key bindings also need
to be updated to reflect these changes.